♻️ chat 타입 리팩토링 (#204)
