+++ src/types/apis/chat.ts
@@ -1,3 +1,5 @@
+import type { Payload } from '@src/types/chat';
+
 export interface MessageReq {
   messageRoomId?: number;
   channelId?: number;
@@ -18,4 +20,38 @@ export interface ReactionReq {
 
 export interface ReplyReq extends MessageReq {
   parentId: string;
+}
+
+export type ChatEventRes = NewMessageEvent | ReactEvent | ReplyEvent;
+
+interface NewMessageEvent extends BaseEvent {
+  eventType: 'NEW_MESSAGE';
+  payload: Payload & {
+    type: 'TEXT' | 'IMAGE';
+  };
+}
+
+interface ReactEvent extends BaseEvent {
+  eventType: 'REACT';
+  payload: {
+    id: string;
+    emoji: string;
+    emojiCount: number;
+    members: number[];
+  };
+}
+
+interface ReplyEvent extends BaseEvent {
+  eventType: 'REPLY';
+  payload: Payload & {
+    parentId: string;
+    parentContent: string;
+    type: 'TEXT' | 'IMAGE';
+  };
+}
+
+interface BaseEvent {
+  eventType: 'NEW_MESSAGE' | 'REACT' | 'REPLY';
+  messageRoomId?: number;
+  channelId?: number;
 }
