+++ src/types/messageRoom.ts
@@ -1,4 +1,4 @@
-import { Payload, Reactions } from "@src/types/chat";
+import type { BasePayload } from '@src/types/chat';
 
 export interface MessageRoom {
   messageRoomId: number;
@@ -9,6 +9,6 @@ export interface MessageRoom {
   recentMessageTime: string;
 }
 
-export interface DM extends Payload {
-  reactions?: Reactions;
+export interface DM extends BasePayload {
+  messageRoomId: number;
 }
